refactor: Change folder name 'Accounts' to 'Decks'
